Create Scene.h
Blocks of code in order to implement a ray tracing algorithm.
We are sending a ray from a pixel through a BVH structure, then calculating contributions of each light composant for the pixel in order to get its final color. 
Then sending several rays for every pixels of the screen.